# openapi/openapi.yaml
components:
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  contact: {}
  description: Backend control-plane API for the adult English voice-room product.
  title: Slogan Voice Room API
  version: 0.0.1
openapi: 3.0.0
paths:
...
  /v1/rooms:
    get:
      operationId: RoomsController_list
      parameters:
        - description: Opaque pagination cursor
          in: query
          name: cursor
          required: false
          schema:
            maxLength: 500
            type: string
        - in: query
          name: limit
          required: false
          schema:
            default: 20
            maximum: 50
            minimum: 1
            type: number
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  items:
                    items:
                      properties:
                        capacity:
                          maximum: 6
                          minimum: 2
                          type: number
                        cefrLevel:
                          enum:
                            - A1
                            - A2
                            - B1
                            - B2
                            - C1
                            - C2
                          type: string
                        endsAt:
                          format: date-time
                          type: string
                        hostDisplayName:
                          type: string
                        hostUserId:
                          format: uuid
                          type: string
                        id:
                          format: uuid
                          type: string
                        memberCount:
                          minimum: 1
                          type: number
                        passwordProtected:
                          type: boolean
                        startedAt:
                          format: date-time
                          type: string
                        topic:
                          type: string
                      required:
                        - id
                        - hostUserId
                        - hostDisplayName
                        - topic
                        - cefrLevel
                        - capacity
                        - memberCount
                        - passwordProtected
                        - startedAt
                        - endsAt
                      type: object
                    type: array
                  nextCursor:
                    description: Opaque cursor for the next page
                    nullable: true
                    type: string
                required:
                  - items
                  - nextCursor
                type: object
          description: ""
        "400":
          content:
            application/json:
              schema:
                properties:
                  code:
                    example: VALIDATION_FAILED
                    type: string
                  details:
                    type: object
                  message:
                    example: Request validation failed
                    type: string
                  requestId:
                    example: 5b27b1e0-f183-46f6-b9e8-1421235a8bb1
                    type: string
                required:
                  - code
                  - message
                type: object
          description: ""
        "401":
          content:
            application/json:
              schema:
                properties:
                  code:
                    example: VALIDATION_FAILED
                    type: string
                  details:
                    type: object
                  message:
                    example: Request validation failed
                    type: string
                  requestId:
                    example: 5b27b1e0-f183-46f6-b9e8-1421235a8bb1
                    type: string
                required:
                  - code
                  - message
                type: object
          description: ""
        "403":
          content:
            application/json:
              schema:
                properties:
                  code:
                    example: VALIDATION_FAILED
                    type: string
                  details:
                    type: object
                  message:
                    example: Request validation failed
                    type: string
                  requestId:
                    example: 5b27b1e0-f183-46f6-b9e8-1421235a8bb1
                    type: string
                required:
                  - code
                  - message
                type: object
          description: ""
      security:
        - bearer: []
      tags:
        - rooms
    post:
      operationId: RoomsController_create
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              properties:
                capacity:
                  maximum: 6
                  minimum: 2
                  type: number
                cefrLevel:
                  enum:
                    - A1
                    - A2
                    - B1
                    - B2
                    - C1
                    - C2
                  type: string
                password:
                  pattern: ^\d{4}$
                  type: string
                  writeOnly: true
                topic:
                  maxLength: 120
                  minLength: 2
                  type: string
              required:
                - topic
                - cefrLevel
                - capacity
              type: object
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                properties:
                  capacity:
                    maximum: 6
                    minimum: 2
                    type: number
                  cefrLevel:
                    enum:
                      - A1
                      - A2
                      - B1
                      - B2
                      - C1
                      - C2
                    type: string
                  currentMembership:
                    allOf:
                      - properties:
                          id:
                            format: uuid
                            type: string
                          joinOrder:
                            minimum: 1
                            type: number
                          joinedAt:
                            format: date-time
                            type: string
                          role:
                            enum:
                              - HOST
                              - MEMBER
                            type: string
                          rulesAcceptedAt:
                            format: date-time
                            type: string
                          rulesVersion:
                            type: string
                          userId:
                            format: uuid
                            type: string
                        required:
                          - id
                          - userId
                          - role
                          - joinOrder
                          - rulesVersion
                          - rulesAcceptedAt
                          - joinedAt
                        type: object
                    nullable: true
                    type: object
                  endsAt:
                    format: date-time
                    type: string
                  hostDisplayName:
                    type: string
                  hostUserId:
                    format: uuid
                    type: string
                  id:
                    format: uuid
                    type: string
                  memberCount:
                    minimum: 1
                    type: number
                  passwordProtected:
                    type: boolean
                  startedAt:
                    format: date-time
                    type: string
                  topic:
                    type: string
                required:
                  - id
                  - hostUserId
                  - hostDisplayName
                  - topic
                  - cefrLevel
                  - capacity
                  - memberCount
                  - passwordProtected
                  - startedAt
                  - endsAt
                  - currentMembership
                type: object
          description: ""
        "400":
          content:
            application/json:
              schema:
                properties:
                  code:
                    example: VALIDATION_FAILED
                    type: string
                  details:
                    type: object
                  message:
                    example: Request validation failed
                    type: string
                  requestId:
                    example: 5b27b1e0-f183-46f6-b9e8-1421235a8bb1
                    type: string
                required:
                  - code
                  - message
                type: object
          description: ""
        "401":
          content:
            application/json:
              schema:
                properties:
                  code:
                    example: VALIDATION_FAILED
                    type: string
                  details:
                    type: object
                  message:
                    example: Request validation failed
                    type: string
                  requestId:
                    example: 5b27b1e0-f183-46f6-b9e8-1421235a8bb1
                    type: string
                required:
                  - code
                  - message
                type: object
          description: ""
        "403":
          content:
            application/json:
              schema:
                properties:
                  code:
                    example: VALIDATION_FAILED
                    type: string
                  details:
                    type: object
                  message:
                    example: Request validation failed
                    type: string
                  requestId:
                    example: 5b27b1e0-f183-46f6-b9e8-1421235a8bb1
                    type: string
                required:
                  - code
                  - message
                type: object
          description: ""
      security:
        - bearer: []
      tags:
        - rooms
  /v1/rooms/{roomId}:
    get:
      operationId: RoomsController_detail
      parameters:
        - in: path
          name: roomId
          required: true
          schema:
            format: uuid
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  capacity:
                    maximum: 6
                    minimum: 2
                    type: number
                  cefrLevel:
                    enum:
                      - A1
                      - A2
                      - B1
                      - B2
                      - C1
                      - C2
                    type: string
                  currentMembership:
                    allOf:
                      - properties:
                          id:
                            format: uuid
                            type: string
                          joinOrder:
                            minimum: 1
                            type: number
                          joinedAt:
                            format: date-time
                            type: string
                          role:
                            enum:
                              - HOST
                              - MEMBER
                            type: string
                          rulesAcceptedAt:
                            format: date-time
                            type: string
                          rulesVersion:
                            type: string
                          userId:
                            format: uuid
                            type: string
                        required:
                          - id
                          - userId
                          - role
                          - joinOrder
                          - rulesVersion
                          - rulesAcceptedAt
                          - joinedAt
                        type: object
                    nullable: true
                    type: object
                  endsAt:
                    format: date-time
                    type: string
                  hostDisplayName:
                    type: string
                  hostUserId:
                    format: uuid
                    type: string
                  id:
                    format: uuid
                    type: string
                  memberCount:
                    minimum: 1
                    type: number
                  passwordProtected:
                    type: boolean
                  startedAt:
                    format: date-time
                    type: string
                  topic:
                    type: string
                required:
                  - id
                  - hostUserId
                  - hostDisplayName
                  - topic
                  - cefrLevel
                  - capacity
                  - memberCount
                  - passwordProtected
                  - startedAt
                  - endsAt
                  - currentMembership
                type: object
          description: ""
        "401":
          content:
            application/json:
              schema:
                properties:
                  code:
                    example: VALIDATION_FAILED
                    type: string
                  details:
                    type: object
                  message:
                    example: Request validation failed
                    type: string
                  requestId:
                    example: 5b27b1e0-f183-46f6-b9e8-1421235a8bb1
                    type: string
                required:
                  - code
                  - message
                type: object
          description: ""
        "403":
          content:
            application/json:
              schema:
                properties:
                  code:
                    example: VALIDATION_FAILED
                    type: string
                  details:
                    type: object
                  message:
                    example: Request validation failed
                    type: string
                  requestId:
                    example: 5b27b1e0-f183-46f6-b9e8-1421235a8bb1
                    type: string
                required:
                  - code
                  - message
                type: object
          description: ""
        "404":
          content:
            application/json:
              schema:
                properties:
                  code:
                    example: VALIDATION_FAILED
                    type: string
                  details:
                    type: object
                  message:
                    example: Request validation failed
                    type: string
                  requestId:
                    example: 5b27b1e0-f183-46f6-b9e8-1421235a8bb1
                    type: string
                required:
                  - code
                  - message
                type: object
          description: ""
        "409":
          content:
            application/json:
              schema:
                properties:
                  code:
                    example: VALIDATION_FAILED
                    type: string
                  details:
                    type: object
                  message:
                    example: Request validation failed
                    type: string
                  requestId:
                    example: 5b27b1e0-f183-46f6-b9e8-1421235a8bb1
                    type: string
                required:
                  - code
                  - message
                type: object
          description: ""
      security:
        - bearer: []
      tags:
        - rooms
  /v1/rooms/{roomId}/memberships:
    post:
      operationId: RoomsController_join
      parameters:
        - in: path
          name: roomId
          required: true
          schema:
            format: uuid
            type: string
      requestBody:
        content:
          application/json:
            schema:
              properties:
                password:
                  pattern: ^\d{4}$
                  type: string
                  writeOnly: true
                rulesAccepted:
                  description: Whether the current room rules were actively accepted
                  type: boolean
              required:
                - rulesAccepted
              type: object
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                properties:
                  capacity:
                    maximum: 6
                    minimum: 2
                    type: number
                  cefrLevel:
                    enum:
                      - A1
                      - A2
                      - B1
                      - B2
                      - C1
                      - C2
                    type: string
                  currentMembership:
                    allOf:
                      - properties:
                          id:
                            format: uuid
                            type: string
                          joinOrder:
                            minimum: 1
                            type: number
                          joinedAt:
                            format: date-time
                            type: string
                          role:
                            enum:
                              - HOST
                              - MEMBER
                            type: string
                          rulesAcceptedAt:
                            format: date-time
                            type: string
                          rulesVersion:
                            type: string
                          userId:
                            format: uuid
                            type: string
                        required:
                          - id
                          - userId
                          - role
                          - joinOrder
                          - rulesVersion
                          - rulesAcceptedAt
                          - joinedAt
                        type: object
                    nullable: true
                    type: object
                  endsAt:
                    format: date-time
                    type: string
                  hostDisplayName:
                    type: string
                  hostUserId:
                    format: uuid
                    type: string
                  id:
                    format: uuid
                    type: string
                  memberCount:
                    minimum: 1
                    type: number
                  passwordProtected:
                    type: boolean
                  startedAt:
                    format: date-time
                    type: string
                  topic:
                    type: string
                required:
                  - id
                  - hostUserId
                  - hostDisplayName
                  - topic
                  - cefrLevel
                  - capacity
                  - memberCount
                  - passwordProtected
                  - startedAt
                  - endsAt
                  - currentMembership
                type: object
          description: ""
        "400":
          content:
            application/json:
              schema:
                properties:
                  code:
                    example: VALIDATION_FAILED
                    type: string
                  details:
                    type: object
                  message:
                    example: Request validation failed
                    type: string
                  requestId:
                    example: 5b27b1e0-f183-46f6-b9e8-1421235a8bb1
                    type: string
                required:
                  - code
                  - message
                type: object
          description: ""
        "401":
          content:
            application/json:
              schema:
                properties:
                  code:
                    example: VALIDATION_FAILED
                    type: string
                  details:
                    type: object
                  message:
                    example: Request validation failed
                    type: string
                  requestId:
                    example: 5b27b1e0-f183-46f6-b9e8-1421235a8bb1
                    type: string
                required:
                  - code
                  - message
                type: object
          description: ""
        "403":
          content:
            application/json:
              schema:
                properties:
                  code:
                    example: VALIDATION_FAILED
                    type: string
                  details:
                    type: object
                  message:
                    example: Request validation failed
                    type: string
                  requestId:
                    example: 5b27b1e0-f183-46f6-b9e8-1421235a8bb1
                    type: string
                required:
                  - code
                  - message
                type: object
          description: ""
        "404":
          content:
            application/json:
              schema:
                properties:
                  code:
                    example: VALIDATION_FAILED
                    type: string
                  details:
                    type: object
                  message:
                    example: Request validation failed
                    type: string
                  requestId:
                    example: 5b27b1e0-f183-46f6-b9e8-1421235a8bb1
                    type: string
                required:
                  - code
                  - message
                type: object
          description: ""
        "409":
          content:
            application/json:
              schema:
                properties:
                  code:
                    example: VALIDATION_FAILED
                    type: string
                  details:
                    type: object
                  message:
                    example: Request validation failed
                    type: string
                  requestId:
                    example: 5b27b1e0-f183-46f6-b9e8-1421235a8bb1
                    type: string
                required:
                  - code
                  - message
                type: object
          description: ""
      security:
        - bearer: []
      tags:
        - rooms
servers: []
...